pyproject.toml
[project]
name = "ralph"
version = "0.1.0"
description = "Deterministic agentic coding loop using Claude Agent SDK"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Ralph Team", email = "ralph@example.com" }
]
dependencies = [
    "claude-agent-sdk>=0.1.21",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "aiofiles>=24.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-aiofiles>=24.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
ralph-agent = "ralph.cli:app"

[project.urls]
Homepage = "https://github.com/CipherScout/Ralph"
Documentation = "https://github.com/CipherScout/Ralph/tree/main/docs"
Repository = "https://github.com/CipherScout/Ralph.git"
Issues = "https://github.com/CipherScout/Ralph/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ralph"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
disable_error_code = ["union-attr", "index", "operator", "comparison-overlap", "no-untyped-def", "method-assign"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
